refactor(framework): rename createBlockMethod to createBlockRequest
The helper does not create a block method, it creates the request function that
block methods are built with - the JSDoc had to say so to correct the name. The
new one matches what it returns and everything around it: BlockRequest,
BlockRequestConfig, BlockRequestError. The module follows.

Nothing has been released under the old name, so there is no alias to keep.
