+++ packages/blocks/billing/invoice-list/src/sdk/invoice-list.ts
@@ -1,5 +1,5 @@
 import { AppHeaders } from '@o2s/framework/headers';
-import { Sdk, createBlockMethod } from '@o2s/framework/sdk';
+import { Sdk, createBlockRequest } from '@o2s/framework/sdk';
 
 import { Model, Request } from '../api-harmonization/invoice-list.client';
 import { URL } from '../api-harmonization/invoice-list.url';
@@ -7,7 +7,7 @@ import { URL } from '../api-harmonization/invoice-list.url';
 const API_URL = URL;
 
 export const invoiceList = (sdk: Sdk) => {
-    const request = createBlockMethod(sdk);
+    const request = createBlockRequest(sdk);
 
     return {
         blocks: {
