+++ packages/blocks/billing/payments-history/src/sdk/payments-history.ts
@@ -1,5 +1,5 @@
 import { AppHeaders } from '@o2s/framework/headers';
-import { Sdk, createBlockMethod } from '@o2s/framework/sdk';
+import { Sdk, createBlockRequest } from '@o2s/framework/sdk';
 
 import { Model, Request } from '../api-harmonization/payments-history.client';
 import { URL } from '../api-harmonization/payments-history.url';
@@ -7,7 +7,7 @@ import { URL } from '../api-harmonization/payments-history.url';
 const API_URL = URL;
 
 export const paymentsHistory = (sdk: Sdk) => {
-    const request = createBlockMethod(sdk);
+    const request = createBlockRequest(sdk);
 
     return {
         blocks: {
