+++ packages/blocks/billing/payments-summary/src/sdk/payments-summary.ts
@@ -1,5 +1,5 @@
 import { AppHeaders } from '@o2s/framework/headers';
-import { Sdk, createBlockMethod } from '@o2s/framework/sdk';
+import { Sdk, createBlockRequest } from '@o2s/framework/sdk';
 
 import { Model, Request } from '../api-harmonization/payments-summary.client';
 import { URL } from '../api-harmonization/payments-summary.url';
@@ -7,7 +7,7 @@ import { URL } from '../api-harmonization/payments-summary.url';
 const API_URL = URL;
 
 export const paymentsSummary = (sdk: Sdk) => {
-    const request = createBlockMethod(sdk);
+    const request = createBlockRequest(sdk);
 
     return {
         blocks: {
