+++ packages/blocks/content/bento-grid/src/sdk/bento-grid.ts
@@ -1,5 +1,5 @@
 import { AppHeaders } from '@o2s/framework/headers';
-import { Sdk, createBlockMethod } from '@o2s/framework/sdk';
+import { Sdk, createBlockRequest } from '@o2s/framework/sdk';
 
 import { Model, Request } from '../api-harmonization/bento-grid.client';
 import { URL } from '../api-harmonization/bento-grid.url';
@@ -7,7 +7,7 @@ import { URL } from '../api-harmonization/bento-grid.url';
 const API_URL = URL;
 
 export const bentoGrid = (sdk: Sdk) => {
-    const request = createBlockMethod(sdk);
+    const request = createBlockRequest(sdk);
 
     return {
         blocks: {
