+++ packages/blocks/content/feature-section-grid/src/sdk/feature-section-grid.ts
@@ -1,5 +1,5 @@
 import { AppHeaders } from '@o2s/framework/headers';
-import { Sdk, createBlockMethod } from '@o2s/framework/sdk';
+import { Sdk, createBlockRequest } from '@o2s/framework/sdk';
 
 import { Model, Request } from '../api-harmonization/feature-section-grid.client';
 import { URL } from '../api-harmonization/feature-section-grid.url';
@@ -7,7 +7,7 @@ import { URL } from '../api-harmonization/feature-section-grid.url';
 const API_URL = URL;
 
 export const featureSectionGrid = (sdk: Sdk) => {
-    const request = createBlockMethod(sdk);
+    const request = createBlockRequest(sdk);
 
     return {
         blocks: {
