+++ packages/blocks/content/quick-links/src/sdk/quick-links.ts
@@ -1,5 +1,5 @@
 import { AppHeaders } from '@o2s/framework/headers';
-import { Sdk, createBlockMethod } from '@o2s/framework/sdk';
+import { Sdk, createBlockRequest } from '@o2s/framework/sdk';
 
 import { Model, Request } from '../api-harmonization/quick-links.client';
 import { URL } from '../api-harmonization/quick-links.url';
@@ -7,7 +7,7 @@ import { URL } from '../api-harmonization/quick-links.url';
 const API_URL = URL;
 
 export const quickLinks = (sdk: Sdk) => {
-    const request = createBlockMethod(sdk);
+    const request = createBlockRequest(sdk);
 
     return {
         blocks: {
