+++ packages/blocks/forms/surveyjs-form/src/sdk/surveyjs.ts
@@ -1,5 +1,5 @@
 import { AppHeaders } from '@o2s/framework/headers';
-import { Sdk, createBlockMethod } from '@o2s/framework/sdk';
+import { Sdk, createBlockRequest } from '@o2s/framework/sdk';
 
 import { Model, Request } from '../api-harmonization/surveyjs.client';
 import { URL } from '../api-harmonization/surveyjs.url';
@@ -7,7 +7,7 @@ import { URL } from '../api-harmonization/surveyjs.url';
 const API_URL = URL;
 
 export const surveyjs = (sdk: Sdk) => {
-    const request = createBlockMethod(sdk);
+    const request = createBlockRequest(sdk);
 
     return {
         blocks: {
