+++ packages/blocks/knowledge-base/article/src/sdk/article.ts
@@ -1,5 +1,5 @@
 import { AppHeaders } from '@o2s/framework/headers';
-import { Sdk, createBlockMethod } from '@o2s/framework/sdk';
+import { Sdk, createBlockRequest } from '@o2s/framework/sdk';
 
 import { Model, Request } from '../api-harmonization/article.client';
 import { URL } from '../api-harmonization/article.url';
@@ -7,7 +7,7 @@ import { URL } from '../api-harmonization/article.url';
 const API_URL = URL;
 
 export const article = (sdk: Sdk) => {
-    const request = createBlockMethod(sdk);
+    const request = createBlockRequest(sdk);
 
     return {
         blocks: {
