+++ packages/blocks/knowledge-base/category-list/src/sdk/category-list.ts
@@ -1,5 +1,5 @@
 import { AppHeaders } from '@o2s/framework/headers';
-import { Sdk, createBlockMethod } from '@o2s/framework/sdk';
+import { Sdk, createBlockRequest } from '@o2s/framework/sdk';
 
 import { Model, Request } from '../api-harmonization/category-list.client';
 import { URL } from '../api-harmonization/category-list.url';
@@ -7,7 +7,7 @@ import { URL } from '../api-harmonization/category-list.url';
 const API_URL = URL;
 
 export const categoryList = (sdk: Sdk) => {
-    const request = createBlockMethod(sdk);
+    const request = createBlockRequest(sdk);
 
     return {
         blocks: {
