+++ packages/blocks/orders/order-details/src/sdk/order-details.ts
@@ -1,5 +1,5 @@
 import { AppHeaders } from '@o2s/framework/headers';
-import { Sdk, createBlockMethod } from '@o2s/framework/sdk';
+import { Sdk, createBlockRequest } from '@o2s/framework/sdk';
 
 import { Model, Request } from '../api-harmonization/order-details.client';
 import { URL } from '../api-harmonization/order-details.url';
@@ -7,7 +7,7 @@ import { URL } from '../api-harmonization/order-details.url';
 const API_URL = URL;
 
 export const orderDetails = (sdk: Sdk) => {
-    const request = createBlockMethod(sdk);
+    const request = createBlockRequest(sdk);
 
     return {
         blocks: {
