+++ packages/blocks/orders/order-list/src/sdk/order-list.ts
@@ -1,11 +1,11 @@
 import { AppHeaders } from '@o2s/framework/headers';
-import { Sdk, createBlockMethod } from '@o2s/framework/sdk';
+import { Sdk, createBlockRequest } from '@o2s/framework/sdk';
 
 import { Model, Request } from '../api-harmonization/order-list.client';
 import { URL } from '../api-harmonization/order-list.url';
 
 export const orderList = (sdk: Sdk) => {
-    const request = createBlockMethod(sdk);
+    const request = createBlockRequest(sdk);
 
     return {
         blocks: {
