+++ packages/blocks/orders/orders-summary/src/sdk/orders-summary.ts
@@ -1,5 +1,5 @@
 import { AppHeaders } from '@o2s/framework/headers';
-import { Sdk, createBlockMethod } from '@o2s/framework/sdk';
+import { Sdk, createBlockRequest } from '@o2s/framework/sdk';
 
 import { Model, Request } from '../api-harmonization/orders-summary.client';
 import { URL } from '../api-harmonization/orders-summary.url';
@@ -7,7 +7,7 @@ import { URL } from '../api-harmonization/orders-summary.url';
 const API_URL = URL;
 
 export const ordersSummary = (sdk: Sdk) => {
-    const request = createBlockMethod(sdk);
+    const request = createBlockRequest(sdk);
 
     return {
         blocks: {
