+++ packages/blocks/products/product-list/src/sdk/product-list.ts
@@ -1,6 +1,6 @@
 import { AppHeaders } from '@o2s/framework/headers';
 import { Carts } from '@o2s/framework/modules';
-import { Sdk, createBlockMethod } from '@o2s/framework/sdk';
+import { Sdk, createBlockRequest } from '@o2s/framework/sdk';
 
 import { Model, Request } from '../api-harmonization/product-list.client';
 import { URL } from '../api-harmonization/product-list.url';
@@ -9,7 +9,7 @@ const API_URL = URL;
 const CARTS_API_URL = '/carts';
 
 export const productList = (sdk: Sdk) => {
-    const request = createBlockMethod(sdk);
+    const request = createBlockRequest(sdk);
 
     return {
         blocks: {
