+++ packages/modules/surveyjs/src/sdk/surveyjs.ts
@@ -1,12 +1,12 @@
 import { AppHeaders } from '@o2s/framework/headers';
-import { Sdk, createBlockMethod } from '@o2s/framework/sdk';
+import { Sdk, createBlockRequest } from '@o2s/framework/sdk';
 
 import { Model, Request, URL } from '../api-harmonization/surveyjs.client';
 
 const API_URL = URL;
 
 export const surveyjs = (sdk: Sdk) => {
-    const request = createBlockMethod(sdk);
+    const request = createBlockRequest(sdk);
 
     return {
         modules: {
